requirements.txt
boto3==1.16.19
psycopg2-binary==2.8.5
